+++ web/apps/admin/app.py
@@ -6,7 +6,8 @@ from zoom.apps import App
 
 app = App()
 app.menu = (
-    'Overview', 'Apps', 'Users', 'Groups', 'Mail',
+    'Overview', 'Apps', 'Users', 'Groups', 
+    'Database', 'Mail',
     'Requests', 'Activity', 'Errors', 'Audit',
     'Configuration', 'About'
 )